requirements.txt
# FastAPI和相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库相关
sqlalchemy==2.0.23
pymysql==1.1.0
mysql-connector-python==8.2.0
alembic==1.12.1

# 数据验证和设置
pydantic==2.5.0
pydantic-settings==2.1.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP客户端（用于大模型API调用）
httpx==0.25.2

# 大模型API客户端
openai==1.3.0

# 其他工具
python-dotenv==1.0.0